% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat.R
\name{cat_ftable}
\alias{cat_ftable}
\title{Compute frequency table}
\usage{
cat_ftable(data, var)
}
\arguments{
\item{data}{A (tidy) dataset}

\item{var}{Variable to describe}
}
\value{
A tibble of the frequency table
}
\description{
Computes frequency table for one categorial variable, including relative,
valid, and cumulative frequencies.
}
