% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat.R
\name{cat_xtable}
\alias{cat_xtable}
\title{Compute contigency table}
\usage{
cat_xtable(data, col.var, ..., percentages = FALSE, chisq = FALSE)
}
\arguments{
\item{data}{A (tidy) dataset}

\item{col.var}{Independent (column) variable.}

\item{...}{Dependt (row) variable(s).}

\item{percentages}{Logical indicating whether to output column-wise
percentages instead of absolute values. Defaults to FALSE.}

\item{chisq}{Logical indicating whether a Chi² test should be computed. Test
will be reported via message(). Defaults to FALSE.}
}
\value{
A tibble with the contingency table.
}
\description{
Computes contingency table for one independent (column) variable and one
or more dependent (row) variables. Optionally computes Chi² test.
}
