% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest.R
\name{cohensD}
\alias{cohensD}
\title{Calculate Cohen's d}
\usage{
cohensD(M1, M2, SD1, SD2)
}
\arguments{
\item{M1}{Mean of Group 1}

\item{M2}{Mean of Group 2}

\item{SD1}{Standard deviation of Group 1}

\item{SD2}{Standard deviation of Group 2}
}
\value{
Absolute value of Cohen's d
}
\description{
Calculates Cohen's d for a t-Test
}
