% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{color_scale}
\alias{color_scale}
\title{Color scales}
\usage{
color_scale(scale = "cb")
}
\arguments{
\item{scale}{Name of the color scale. Defaults to "cb".}
}
\value{
A vector containing the color scale (hex values).
}
\description{
A variety of discreet color scales to be used in ggplot as a value in
scale_color_manual().
Currently, the following scales are supported:
- "cb": Colorblind-friendly scale.
- "cb2": Alternative colorblind-friendly scale.
}
\references{
\url{http://jfly.iam.u-tokyo.ac.jp/color/}
}
