% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.R
\name{correlate}
\alias{correlate}
\title{Compute correlation matrix}
\usage{
correlate(data, ..., type = "pearson")
}
\arguments{
\item{data}{A (tidy) dataset.}

\item{...}{Variables to correlate with eachother.}

\item{type}{Type of correlation coefficient. Either "pearson" or "spearman".}
}
\value{
A tibble containing correlation coefficients.
}
\description{
Computes correlations coefficients with p-values for all specified variables.
}
