% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe}
\alias{describe}
\title{Describe variables}
\usage{
describe(data, ..., na.rm = T)
}
\arguments{
\item{data}{A (tidy) dataset.}

\item{...}{Variables to describe (column names)}

\item{na.rm}{a logical value indicating whether NA values should be stripped
before the computation proceeds. Defaults to TRUE.}
}
\value{
A tibble containing descriptive statistics for the selected
    variables
}
\description{
Descriptively describe variables by several measures of
central tendency and variability. If no variables are specified,
all numeric (integer or double) variables are described.
}
\examples{
describe(mtcars, disp, hp, mpg)
  describe(iris)

}
