% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe_cat}
\alias{describe_cat}
\title{Describe categorial variable}
\usage{
describe_cat(data, var, sort = F)
}
\arguments{
\item{data}{A (tidy) dataset}

\item{var}{Variable to describe}

\item{sort}{A logical value indicating whether categories should be sorted
descending by highest value. Defaults to FALSE.}
}
\value{
A tibble containing absolute frequencies, percentages and cumulative
    percentages.
}
\description{
Compute frequency table for a categorial variable (includes percentages
 and cumulative percentages).
}
\examples{
describe_cat(mtcars, cyl)

}
