% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irr.R
\name{irr.matrix}
\alias{irr.matrix}
\title{Create codings matrix}
\usage{
irr.matrix(data, unitVar, coderVar, testVar)
}
\arguments{
\item{data}{A (tidy) dataset}

\item{unitVar}{A unit (e.g., an article) identifier variable.}

\item{coderVar}{A coder identifier variable.}

\item{testVar}{The variable to be tested.}
}
\value{
A coders x codings matrix.
}
\description{
Creates a coders x codings matrix for a test variable that can be passed to
irr functions.
}
