% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irr.R
\name{irr.test}
\alias{irr.test}
\title{Compute interreliability test}
\usage{
irr.test(data, unitVar, coderVar, testVar, method = "nominal")
}
\arguments{
\item{data}{A (tidy) dataset}

\item{unitVar}{A unit (e.g., an article) identifier variable.}

\item{coderVar}{A coder identifier variable.}

\item{testVar}{The variable to be tested.}

\item{method}{Data level of the test variable ("nominal", "ordinal",
"interval" or "ratio"). Defaults to "nominal".}
}
\value{
A tibble containing various interreliability measures.
}
\description{
Computes an interreliability test for a specific test variable.
}
