% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{label_text}
\alias{label_text}
\title{Label text size multiplier}
\usage{
label_text(size = 12)
}
\arguments{
\item{size}{Text size in default units to be used in geom_text().}
}
\value{
The desired text size in points.
}
\description{
Multiplier for text size in geom_text() to correspond with element_text().
}
\references{
\url{https://stackoverflow.com/questions/17311917/ggplot2-the-unit-of-size}
}
