% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{summary}
\alias{summary}
\title{Provide summary functions}
\usage{
summary(var, na.rm)
}
\arguments{
\item{var}{Variable that holds values to summarize}

\item{na.rm}{a logical value indicating whether NA values should be stripped
before the computation proceeds, passed on from the describe functions.}
}
\value{
A list of summary functions to be used in summarise().
}
\description{
Returns summary functions for describe functions
}
