% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{theme_ju}
\alias{theme_ju}
\title{Plot theme}
\usage{
theme_ju(base_size = 12, base_family = "sans")
}
\arguments{
\item{base_size}{Base size for text elements. Defaults to 12.}

\item{base_family}{Base family for text elements. Defaults to "sans".}
}
\value{
Theme to be added to ggplot2 object.
}
\description{
ggplot2 theme based on theme_minimal()
}
