% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest.R
\name{ttest}
\alias{ttest}
\title{Compute t-Test}
\usage{
ttest(data, groupVar, ...)
}
\arguments{
\item{data}{A (tidy) dataset.}

\item{groupVar}{Grouping variable}

\item{...}{Test variables.}
}
\value{
A tibble containing test statistics for all test variables.
}
\description{
Computes t-Tests for one grouping variable and one or more test variables.
}
